fix %EXT% not work in plugin , https://github.com/chainreactors/spray/issues/63
