+++ internal/types.go
@@ -14,13 +14,13 @@ type Origin struct {
 	sum int
 }
 
-func (o *Origin) InitWorder(fns []func(string) []string) (*words.Worder, error) {
+func (o *Origin) InitWorder(fns []words.WordFunc) (*words.Worder, error) {
 	var worder *words.Worder
 	wl, err := pkg.LoadWordlist(o.Word, o.Dictionaries)
 	if err != nil {
 		return nil, err
 	}
-	worder = words.NewWorder(wl)
+	worder = words.NewWorderWithList(wl)
 	worder.Fns = fns
 	rules, err := pkg.LoadRuleWithFiles(o.RuleFiles, o.RuleFilter)
 	if err != nil {
